third_party/blink/web_tests/external/wpt/css/css-flexbox/intrinsic-size/row-008-expected.txt

This is a testharness.js-based test.
PASS .min-width-flexbox 1
FAIL .min-width-flexbox 2 assert_equals: 
<div class="min-width-flexbox" data-expected-width="310"><div data-expected-width="200" style="flex: 0 1 200px; width: 500px;"><div></div></div><div data-expected-width="101" style="flex: 0.1 0 100px; width: 200px;"><div></div></div></div>
width expected 310 but got 400
FAIL .min-width-flexbox 3 assert_equals: 
<div class="min-width-flexbox" data-expected-width="330"><div data-expected-width="203" style="flex: 0.1 1 200px; width: 500px;"><div></div></div><div data-expected-width="103" style="flex: 0.1 0 100px; width: 200px;"><div></div></div></div>
width expected 330 but got 700
FAIL .min-width-flexbox 4 assert_equals: 
<div class="min-width-flexbox" data-expected-width="420"><div data-expected-width="248" style="flex: 0.4 1 200px; width: 500px;"><div></div></div><div data-expected-width="148" style="flex: 0.4 0 100px; width: 200px;"><div></div></div></div>
width expected 420 but got 700
FAIL .min-width-flexbox 5 assert_equals: 
<div class="min-width-flexbox" data-expected-width="450"><div data-expected-width="275" style="flex: 0.5 1 200px; width: 500px;"><div></div></div><div data-expected-width="175" style="flex: 0.5 0 100px; width: 200px;"><div></div></div></div>
width expected 450 but got 700
FAIL .min-width-flexbox 6 assert_equals: 
<div class="min-width-flexbox" data-expected-width="637.5"><div data-expected-width="368.75" style="flex: 0.75 1 200px; width: 500px;"><div></div></div><div data-expected-width="268.75" style="flex: 0.75 0 100px; width: 200px;"><div></div></div></div>
width expected 637.5 but got 700
FAIL .min-width-flexbox 7 assert_equals: 
<div class="min-width-flexbox" data-expected-width="900"><div data-expected-width="500" style="flex: 1 1 200px; width: 500px;"><div></div></div><div data-expected-width="400" style="flex: 1 0 100px; width: 200px;"><div></div></div></div>
width expected 900 but got 700
PASS .min-width-flexbox 8
FAIL .min-width-flexbox 9 assert_equals: 
<div class="min-width-flexbox" data-expected-width="405"><div data-expected-width="205" style="flex: 0.1 1 200px; width: 500px;"><div></div></div><div data-expected-width="200" style="flex: 2 0 100px; width: 200px;"><div></div></div></div>
width expected 405 but got 700
FAIL .min-width-flexbox 10 assert_equals: 
<div class="min-width-flexbox" data-expected-width="432"><div data-expected-width="212" style="flex: 0.2 1 200px; width: 500px;"><div></div></div><div data-expected-width="220" style="flex: 2 0 100px; width: 200px;"><div></div></div></div>
width expected 432 but got 700
FAIL .min-width-flexbox 11 assert_equals: 
<div class="min-width-flexbox" data-expected-width="900"><div data-expected-width="500" style="flex: 2 1 200px; width: 500px;"><div></div></div><div data-expected-width="400" style="flex: 2 0 100px; width: 200px;"><div></div></div></div>
width expected 900 but got 700
Harness: the test ran to completion.
